requirements.txt
# Once published to PyPI, replace with: openfaas-sdk
openfaas-sdk @ git+https://github.com/openfaas/python-sdk.git
